CB-4489 - fixing network creation and deletion in env service

+++ cloud-aws/src/main/java/com/sequenceiq/cloudbreak/cloud/aws/AwsNetworkConnector.java
@@ -210,18 +210,20 @@ public class AwsNetworkConnector implements NetworkConnector {
 
     @Override
     public void deleteNetworkWithSubnets(NetworkDeletionRequest networkDeletionRequest) {
-        AwsCredentialView credentialView = new AwsCredentialView(networkDeletionRequest.getCloudCredential());
-        AmazonCloudFormationRetryClient cloudFormationRetryClient = getCloudFormationRetryClient(credentialView, networkDeletionRequest.getRegion());
-        DeleteStackRequest deleteStackRequest = new DeleteStackRequest();
-        deleteStackRequest.setStackName(networkDeletionRequest.getStackName());
-        cloudFormationRetryClient.deleteStack(deleteStackRequest);
+        if (!networkDeletionRequest.isExisting()) {
+            AwsCredentialView credentialView = new AwsCredentialView(networkDeletionRequest.getCloudCredential());
+            AmazonCloudFormationRetryClient cloudFormationRetryClient = getCloudFormationRetryClient(credentialView, networkDeletionRequest.getRegion());
+            DeleteStackRequest deleteStackRequest = new DeleteStackRequest();
+            deleteStackRequest.setStackName(networkDeletionRequest.getStackName());
+            cloudFormationRetryClient.deleteStack(deleteStackRequest);
 
-        LOGGER.debug("CloudFormation stack deletion request sent with stack name: '{}' ", networkDeletionRequest.getStackName());
-        PollTask<Boolean> pollTask = getDeleteNetworkPollTask(credentialView, networkDeletionRequest.getRegion(), networkDeletionRequest.getStackName());
-        try {
-            awsBackoffSyncPollingScheduler.schedule(pollTask);
-        } catch (RuntimeException | InterruptedException | ExecutionException | TimeoutException e) {
-            throw new CloudConnectorException(e.getMessage(), e);
+            LOGGER.debug("CloudFormation stack deletion request sent with stack name: '{}' ", networkDeletionRequest.getStackName());
+            PollTask<Boolean> pollTask = getDeleteNetworkPollTask(credentialView, networkDeletionRequest.getRegion(), networkDeletionRequest.getStackName());
+            try {
+                awsBackoffSyncPollingScheduler.schedule(pollTask);
+            } catch (RuntimeException | InterruptedException | ExecutionException | TimeoutException e) {
+                throw new CloudConnectorException(e.getMessage(), e);
+            }
         }
     }
